Replace STREAMBOX_APP_ABSPATH with just APP
I thought it might be possible that instruments could run the app even
if the app wasn't installed on the iPhone, but rather instead you give
instruments the absolute path on your macbook to the app (i.e.
/Users/demo/Downloads/iPhone/streambox-app.3.0.14/meproiphone.app) and
instruments would somehow copy the app to the iPhone and run it
there (or something), but this doesn't work.

The app (meproiphone.app) needs to be installed on the iPhone, so there
is no need to feed instruments the absolue path to the app.  Its
actually misleading to do so if instruments requires that the app be
installed because you could feed instruments the absolue path to
v3.0.14 (for example), but the meproiphone.app installed on the iPhone
is a different version.

So, change the instruments parameter from the absolute path back to just
meproiphone.app.